==> R_scripts/03_EDA_02_DAVID analysis.Rmd <==
---
title: "DAVID analysis"
output: html_notebook
editor_options: 
  chunk_output_type: inline
---

### Packages

```{r}
library(dplyr)
library(styler)
library(ggpubr)
```

# Start here if you have df in Data \> For_DAVID folder. Skip to step 1

```{r}
load("C:/Users/1aria/Desktop/rep/Data/Output_data/For_DAVID/GO_full_df.Rda")   # most complete df, probably unnecessary
load("C:/Users/1aria/Desktop/rep/Data/Output_data/For_DAVID/top_10_unique_GO_sel.Rda")
load("C:/Users/1aria/Desktop/rep/Data/Output_data/For_DAVID/top_10_shared_genes_GO_sel.Rda")


# top 10 selected by lowest FDR, 'sel' because only relevant columns selected

```

# Start here if you don't have the datasets

## Step 0 - find the DAVID terms

### Step 0 - Load df

```{r}
wide_matrix_t_means <- load("C:/Users/1aria/Desktop/rep/Data/Output_data/wide_matrix_t_means.Rda")
```

### Step 0 - Get list of shared genes to submit to DAVID

Done this before, so only get names of 1786 genes/proteins

```{r}
# 1. get list of all genes in at least 2 sets
shared_genes <- wide_matrix_t_means %>%
  tibble::rownames_to_column("Gene_name") %>%
  select(Gene_name) # 1786 genes

write.table(shared_genes,
  file = "C:/Users/1aria/Desktop/rep/Output/Gene_lists/shared_genes.txt",
  col.names = FALSE,
  row.names = FALSE,
  quote = FALSE
)


```

### Step 0 - Run DAVID analysis directly on web. Save txt files.

We had already done the searches for the genes unique to set1. "rep/Data/DAVID/DAVID_xx_shared_genes" "rep/Data/DAVID/DAVID_unique_to_1\_BP..."

### Step 0 - Turn txt files into df

```{r}
# Biological processes
shared_genes_BP <- read.delim("C:/Users/1aria/Desktop/rep/Data/DAVID/DAVID_BP_shared_genes.txt", header = TRUE, sep = "\t") %>% filter(!PValue == 0) 
unique_BP <- read.delim("C:/Users/1aria/Desktop/rep/Data/DAVID/DAVID_unique_to_1_GO_Biological_Process.txt", header = TRUE, sep = "\t") %>% filter(!PValue == 0)

# Molecular function
shared_genes_MF <- read.delim("C:/Users/1aria/Desktop/rep/Data/DAVID/DAVID_MF_shared_genes.txt", header = TRUE, sep = "\t") %>% filter(!PValue == 0)
unique_MF <- read.delim("C:/Users/1aria/Desktop/rep/Data/DAVID/DAVID_unique_to_1_GO_Molecular_Function.txt", header = TRUE, sep = "\t") %>% filter(!PValue == 0)

# Cellular compartment
shared_genes_CC <- read.delim("C:/Users/1aria/Desktop/rep/Data/DAVID/DAVID_CC_shared_genes.txt", header = TRUE, sep = "\t") %>% filter(!PValue == 0)
unique_CC <- read.delim("C:/Users/1aria/Desktop/rep/Data/DAVID/DAVID_unique_to_1_GO_Cellular_Compartment.txt", header = TRUE, sep = "\t") %>% filter(!PValue == 0)
```

### Step 0 - Merge df into one big GO df.

Columns are GO term, source, FDR value, count (=how many genes belong there)

```{r}
# For shared genes & unique genes, select only relevant columns for barplot

shared_genes_BP_sel <- shared_genes_BP %>%
  select(c("Category", "Count", "Term", "FDR")) %>%
  mutate(GO = "Biological process")
shared_genes_MF_sel <- shared_genes_MF %>%
  select(c("Category", "Count", "Term", "FDR")) %>%
  mutate(GO = "Molecular function")
shared_genes_CC_sel <- shared_genes_CC %>%
  select(c("Category", "Count", "Term", "FDR")) %>%
  mutate(GO = "Cellular compartment")
unique_BP_sel <- unique_BP %>%
  select(c("Category", "Count", "Term", "FDR")) %>%
  mutate(GO = "Biological process")
unique_MF_sel <- unique_MF %>%
  select(c("Category", "Count", "Term", "FDR")) %>%
  mutate(GO = "Molecular function")
unique_CC_sel <- unique_CC %>%
  select(c("Category", "Count", "Term", "FDR")) %>%
  mutate(GO = "Cellular compartment")

shared_genes_GO_sel <- bind_rows(shared_genes_BP_sel, shared_genes_MF_sel, shared_genes_CC_sel) %>% mutate(Source = "Shared")
unique_GO_sel <- bind_rows(unique_BP_sel, unique_MF_sel, unique_CC_sel) %>% mutate(Source = "Unique to set 1")

GO_full_df <- bind_rows(shared_genes_GO_sel, unique_GO_sel)
save(GO_full_df, file = 'C:/Users/1aria/Desktop/rep/Data/Output_data/For_DAVID/GO_full_df.Rda')
```

# Step 1: Plot - GO terms (proteins unique to set 1)

```{r, fig.height = 30, fig.width=21}
top_10_unique_GO_sel <- unique_GO_sel %>%
  group_by(GO) %>%
  arrange(FDR) %>%
  slice(1:10) %>%
  mutate(Term_short = substring(Term, first = 12))
top_10_shared_genes_GO_sel <- shared_genes_GO_sel %>%
  group_by(GO) %>%
  arrange(FDR) %>%
  slice(1:10) %>%
  mutate(Term_short = substring(Term, first = 12))

top_10_unique_BP_sel <- unique_BP_sel %>%
  arrange(FDR) %>%
  slice(1:10) %>%
  mutate(Term_short = substring(Term, first = 12))

save(top_10_unique_GO_sel, file = 'C:/Users/1aria/Desktop/rep/Data/Output_data/For_DAVID/top_10_unique_GO_sel.Rda')
save(top_10_shared_genes_GO_sel, file = 'C:/Users/1aria/Desktop/rep/Data/Output_data/For_DAVID/top_10_shared_genes_GO_sel.Rda')


### Plots

shared_barplot <-
  ggplot(top_10_shared_genes_GO_sel, aes(x = reorder(Term_short, FDR), y = Count)) +
  geom_bar(stat = "identity", fill = "#1380A1") +
  theme_bw() +
  # geom_hline(yintercept = 5, linewidth = 0.5, colour="coral3", lty=2)+   # 5 would be 0.05 significance level of FDR
  labs(title = "GO terms (Common proteins)", tag = "A") +
  theme(
    axis.text.x = element_text(
      angle = 90,
      vjust = 0.5,
      hjust = 1
    ),
    axis.title.x = element_blank(),
    plot.title = element_text(size = 12, face = "bold", hjust = 0.5),
    plot.margin=unit(c(1,1,-0.2,1), "cm"))+
  facet_wrap( ~ GO, scales = "free")

shared_barplot

unique_barplot <-
  ggplot(top_10_unique_GO_sel, aes(x = reorder(Term_short, FDR), y = Count)) +
  geom_bar(stat = "identity", fill = "#1380A1") +
  theme_bw() +
  # geom_hline(yintercept = 5, linewidth = 0.5, colour="coral3", lty=2)+   # 5 would be 0.05 significance level of FDR
  labs(title = "GO terms (Proteins unique to set 1)", tag = "B") +
  theme(
    axis.text.x = element_text(
      angle = 90,
      vjust = 0.5,
      hjust = 1
    ),
    axis.title.x = element_blank(),
    plot.title = element_text(size = 12, face = "bold", hjust = 0.5),
    plot.margin=unit(c(0,1,1,1), "cm"))+
  facet_wrap( ~ GO, scales = "free")

unique_barplot 

panel_GO <- ggarrange(shared_barplot, unique_barplot, nrow = 2)

panel_GO

ggsave(path = "C:/Users/1aria/Desktop/rep/Output/Plots", filename = "Panel_GO_terms.png",panel_GO, width=21, height = 24, units="cm",dpi=200)
```
