==> R_scripts/04_Analysis_01_PCA.Rmd <==
---
title: "R Notebook"
output: html_notebook
editor_options: 
  chunk_output_type: inline
fig_width: 15
fig_height: 17
---

'follow: <https://core.ac.uk/reader/216334901?utm_source=linkout>'

### Where to start?

'Check if you have ungappy_wide_matrix_t\_for_PCA.Rda. If so, you can skip the first 0 steps.'

```{r}
load("C:/Users/1aria/Desktop/rep/Data/Output_data/For_PCA/ungappy_wide_matrix_t_for_PCA.Rda")
load("C:/Users/1aria/Desktop/rep/Data/Output_data/For_PCA/ungappy_wide_matrix_t_means_for_PCA.Rda")
Metatable <- read_excel("C:/Users/1aria/Desktop/rep/Data/Metadata/Metatable.xlsx", sheet = "Cycles")
```

### 0. Load packages

```{r}
library(ggpubr)
library(dichromat)
library(readxl)
library(tidyverse)
library(scales)
library(ggrepel)
```

### 0. Load the protein abundance wide df and metadata - with and without mean of sample 2 and 9

```{r}
load("C:/Users/1aria/Desktop/rep/Data/Output_data/wide_matrix.RDa")
load("C:/Users/1aria/Desktop/rep/Data/Output_data/wide_matrix_t_means.Rda")
Metatable <- read_excel("C:/Users/1aria/Desktop/rep/Data/Metadata/Metatable.xlsx", sheet = "Cycles")
Metatable_means <- read_excel("C:/Users/1aria/Desktop/rep/Data/Metadata/Metatable_means.xlsx", 
    sheet = "w nur")
```

### 0. Process matrices

```{r}
wide_matrix_t <- as.data.frame(t(wide_matrix)) %>% select(-contains(c("IC", 'X1C3')))  # transform + delete ICI

# Remove batch 1
wide_matrix_t <- wide_matrix_t %>% select(-contains(c("NR_13", "NR_14", "NR_15", "NR_16", "NR_17", "NR_18")))
wide_matrix_t_means <- wide_matrix_t_means %>% select(-contains(c("NR_13", "NR_14", "NR_15", "NR_16", "NR_17", "NR_18")))

# Save > Output_data > For_PCA
save(wide_matrix_t, file = "wide_matrix_t_for_PCA.Rda")
save(wide_matrix_t_means, file = "wide_matrix_t_means_for_PCA.Rda")
```

### 0. Eliminate NAs from wide_matrix_t

```{r}
# Preprocessing for PCA: cannot work with gappy matrix! --> drop genes that have Nas
# Still, prcomp works by having samples as rows, so need to transform it after

ungappy_wide_matrix_t <- drop_na(wide_matrix_t)
ungappy_wide_matrix_t_means <- drop_na(wide_matrix_t_means)

ungappy_wide_matrix <- as.data.frame(t(ungappy_wide_matrix_t)) %>% mutate_if(is.character, as.numeric)
ungappy_wide_matrix_means <- as.data.frame(t(ungappy_wide_matrix_t_means)) %>% mutate_if(is.character, as.numeric)

save(ungappy_wide_matrix_t, file = file.path('C:','Users', '1aria', 'Desktop', 'rep', 'Data', 'Output_data', 'For_PCA', 'ungappy_wide_matrix_t_for_PCA.Rda'))

save(ungappy_wide_matrix_t_means, file = file.path('C:','Users', '1aria', 'Desktop', 'rep', 'Data', 'Output_data', 'For_PCA', 'ungappy_wide_matrix_t_means_for_PCA.Rda'))

```

# PCA

### 1. Create a PCA object, scree plots

```{r}
# PCA object
pca <-prcomp(ungappy_wide_matrix_t_means,scale=TRUE,retx=TRUE)  # full df


# Scree plot requirements
frac_var <- function(x) x^2/sum(x^2)
library(scales)

# Scree plot for wide_matrix with means (3x replicates are substituted with one mean value per protein)
scree_means <- pca$sdev %>% 
  as_tibble() %>% 
  frac_var() %>% 
  mutate(Comp = colnames(pca$x)) %>% 
  slice(1:9) %>% 
  ggplot(aes(x=Comp, y = value)) + 
  geom_bar(stat = "identity", fill = "#1380A1") +
  # geom_hline(yintercept = 0.03, linetype=2) +
  ylim(c(0, 100)) +
  ggtitle('Scree plot')+
  geom_text(aes(label=signif(value, digits = 3)*100), nudge_y= 0.01 )+
  xlab("Principal components") +
  # ylab("Variance explained") +
  scale_y_continuous(name = "Variance Explained", breaks = seq(0,0.8,0.1), labels = percent_format(accuracy = 5L)) +
  theme_classic(base_size = 14)

scree_means

ggsave(path = "C:/Users/1aria/Desktop/rep/Output/Plots", filename = "Scree_plot_PCA_with_means.png",scree_means, height=15,width=17,units="cm",dpi=200)

# PCA - full df
results_pca <- pca$x %>%
  as.data.frame() %>%
  rownames_to_column(var = "Sample_name")  %>%
  full_join(Metatable, by = "Sample_name") %>%
  filter(Sample_status %in% c('NR', 'R'))  %>%
  filter(Cycle == 0)

ggplot(results_pca, aes(x=PC1, y=PC2)) + #, color = Sample_status)) +
  geom_point(size = 3,  aes(shape = Sample_status, color= factor(Sample_nr) )) + # alpha = as.integer(as.factor(Cycle))
  geom_vline(xintercept = 0, linetype=2) +
  geom_hline(yintercept = 0, linetype=2) +
  theme_bw() +
  theme(panel.grid.major = element_blank(), 
        panel.grid.minor = element_blank())




```

```{r}
### Scree plot from technical reps script
scree_plot_replicates <- pca$sdev %>%
  as_tibble() %>%
  frac_var() %>%
  mutate(Comp = colnames(pca$x)) %>%
  slice(1:9) %>%
  ggplot(aes(x = Comp, y = value)) +
  geom_bar(stat = "identity", fill = "#4DC5F9") +
  geom_hline(yintercept = 0.03, linetype = 2) +
  # xlab("Principal Components") +
  labs(title = "Scree plot", x = "Principal Components") +
  scale_y_continuous(
    name = "Variance Explained",
    breaks = seq(0, 0.8, 0.1),
    labels = percent_format(accuracy = 5L)
  ) +
  theme_classic(base_size = 14)+
   theme(
    plot.title = element_text(
      size = 15,
      face = 'bold',
      vjust = 0
    ),
    plot.subtitle = element_text(size = 12, vjust = -3),
    # axis.text.x = element_blank(),
    axis.title.x = element_text(size = 10),
    axis.title.y = element_text(size = 10),
    axis.text.y = element_text(size = 8)
  )

scree_plot_replicates
```



```{r}
# All the df for PCA

# all
results_pca <- pca$x %>%
  as.data.frame() %>%
  rownames_to_column(var = "Sample_name")  %>%
  full_join(Metatable, by = "Sample_name") %>%
  filter(Sample_status %in% c('NR', 'R'))


# all, SCR only
results_pca_SCR <- pca$x %>%
  as.data.frame() %>%
  rownames_to_column(var = "Sample_name")  %>%
  full_join(Metatable, by = "Sample_name") %>%
  filter(Sample_status %in% c('NR', 'R'))  %>%
  filter(Cycle == 0)

# mean
results_pca_means <- pca_means$x %>%
  as.data.frame() %>%
  rownames_to_column(var = "Sample_name")  %>%
  full_join(Metatable, by = "Sample_name") %>%
  filter(Sample_status %in% c('NR', 'R')) 

# mean, SCR only
results_pca_means_SCR <- pca_means$x %>%
  as.data.frame() %>%
  rownames_to_column(var = "Sample_name")  %>%
  full_join(Metatable, by = "Sample_name") %>%
  filter(Sample_status %in% c('NR', 'R'))  %>%
  filter(Cycle == 0)
```

```{r}
### PCA - version 1

PCA_w_repl_v1 <- ggplot(results_pca, aes(x=PC1, y=PC2, color = Sample_status)) + 
  geom_point(size = 3, aes(alpha = as.factor(Cycle)))+ # alpha = as.integer(as.factor(Cycle))
  geom_vline(xintercept = 0, linetype=2) +
  geom_hline(yintercept = 0, linetype=2) +
  theme_bw() +
  xlim(c(-60, 60))+
  ylim(c(-50, 50))+
  ggtitle('With replicates')+
  scale_color_manual(values=c('royalblue4','coral3'), labels = c("Non Responder", "Responder"))+
  scale_alpha_manual(values=c(0.1, 0.3, 0.6, 0.8, 1),labels = c("Screen", "C3", "C4", "C5", "EoT"))+
  labs(alpha = "Cycle", color = "Treatment outcome")+
  theme(panel.grid.major = element_blank(), 
        panel.grid.minor = element_blank(),
        legend.position = 'bottom')

PCA_w_repl_v1


PCA_w_repl_SCR <- ggplot(results_pca_SCR, aes(x=PC1, y=PC2)) + #, color = Sample_status)) +
  geom_point(size = 3,  aes(shape = Sample_status, color= factor(Sample_nr) )) + # alpha = as.integer(as.factor(Cycle))
  geom_vline(xintercept = 0, linetype=2) +
  geom_hline(yintercept = 0, linetype=2) +
  theme_bw() +
  xlim(c(-60, 60))+
  ylim(c(-50, 50))+
  theme(panel.grid.major = element_blank(), 
        panel.grid.minor = element_blank(),
        legend.position = 'none')



PCA_means <- ggplot(results_pca_means_SCR, aes(x=PC1, y=PC2)) + #, color = Sample_status)) +
  geom_point(size = 3,  aes(shape = Sample_status, color= factor(Sample_nr) )) + # alpha = as.integer(as.factor(Cycle))
  geom_vline(xintercept = 0, linetype=2) +
  geom_hline(yintercept = 0, linetype=2) +
  theme_bw() +
  xlim(c(-60, 60))+
  ylim(c(-50, 50))+
  theme(panel.grid.major = element_blank(), 
        panel.grid.minor = element_blank(),
        legend.position = 'none')

plot_v1 <- ggarrange(PCA_w_repl_v1, PCA_means_v1, common.legend = TRUE)

annotate_figure(plot_v1, top = text_grob("PCA plot - version 1", 
               color = "#333333", face = "bold", size = 14))
```

```{r}
# Alternative PCA plot

PCA_w_repl_v2 <- ggplot(results_pca, aes(x=PC1, y=PC2, color = as.factor(Sample_nr))) + 
  geom_point(size = 3, aes(shape = as.factor(Sample_status)))+ # alpha = as.integer(as.factor(Cycle))
  geom_vline(xintercept = 0, linetype=2) +
  geom_hline(yintercept = 0, linetype=2) +
  theme_bw() +
  xlim(c(-60, 60))+
  ylim(c(-50, 50))+
  ggtitle('With replicates')+
  labs(shape= "Treatment outcome")+
  geom_text_repel(aes(label = Sample_nr))+
  guides(color=FALSE)+
  scale_shape_manual(values=c(1, 19), labels = c("Non Responder", "Responder"))+
  theme(panel.grid.major = element_blank(), 
        panel.grid.minor = element_blank(),
        legend.position = 'bottom')

PCA_w_repl_v2

PCA_means_v2 <- ggplot(results_pca_means, aes(x=PC1, y=PC2, color = as.factor(Sample_nr))) + 
  geom_point(size = 3, aes(shape = as.factor(Sample_status)))+ # alpha = as.integer(as.factor(Cycle))
  geom_vline(xintercept = 0, linetype=2) +
  geom_hline(yintercept = 0, linetype=2) +
  theme_bw() +
  xlim(c(-60, 60))+
  ylim(c(-50, 50))+
  ggtitle('Means')+
   scale_shape_manual(values=c(1, 19), labels = c("Non Responder", "Responder"))+
  geom_text_repel(aes(label = Sample_nr))+
  labs(shape = "Responder status", color = "Patient #")+
  theme(panel.grid.major = element_blank(), 
        panel.grid.minor = element_blank(),
        legend.position = 'none')

PCA_means_v2




plot_v2 <- ggarrange(PCA_w_repl_v2, PCA_means_v2, common.legend = TRUE)

annotate_figure(plot_v2, top = text_grob("PCA plot - version 2", 
               color = "#333333", face = "bold", size = 15))

```

### Plot1 for layout: show replicates

```{r}
# Version 3: show only the replicates

results_pca_replicates_only <- results_pca %>% filter(Cycle == 0 & (Sample_nr == 2|  Sample_nr == 9) )

PCA_w_repl_only <- ggplot(results_pca_replicates_only, aes(x=PC1, y=PC2, color = as.factor(Sample_status))) + 
  geom_point(size = 3)+  #alpha = as.integer(as.factor(Replicate)
  geom_vline(xintercept = 0, linetype=2, col="grey60") +
  geom_hline(yintercept = 0, linetype=2, col="grey60") +
  theme_bw() +
  xlim(c(-60, 60))+
  ylim(c(-50, 50))+
  ggtitle('Replicates')+
  labs(color= "Treatment outcome")+
  geom_text_repel(aes(label = Sample_nr), show.legend=FALSE)+
  scale_color_manual(values=c('royalblue4', 'coral3'), labels = c("Non Responder", "Responder"))+
  theme(panel.grid.major = element_blank(), 
        panel.grid.minor = element_blank(),
        legend.position = 'bottom'
        )

PCA_w_repl_only

ggsave(path = "C:/Users/1aria/Desktop/rep/Output/Plots", filename = "PCA_highlight_replicates.png",PCA_w_repl_only,height=15,width=17,units="cm",dpi=200)




```

### Plot2 for layout: individual patients, pre-treatment

```{r}

results_pca_SCR_means <- results_pca_means %>% filter(Cycle == 0 & Sample_status %in% c('R', 'NR') )

PCA_means_SCR <- ggplot(results_pca_SCR_means , aes(x=PC1, y=PC2, color = as.factor(Sample_status))) + 
  geom_point(size = 3)+  #alpha = as.integer(as.factor(Replicate)
  geom_vline(xintercept = 0, linetype=2, col="grey60") +
  geom_hline(yintercept = 0, linetype=2, col="grey60") +
  theme_bw() +
  xlim(c(-60, 60))+
  ylim(c(-50, 50))+
  ggtitle('Patients before treatment') +
  labs(color= "Treatment outcome") +
  geom_text_repel(aes(label = Sample_nr), show.legend = FALSE)+
  scale_color_manual(values=c("royalblue4", 'coral3'), labels = c("Non Responder", "Responder")) +
  theme(panel.grid.major = element_blank(), 
        panel.grid.minor = element_blank(),
        legend.position = 'none'
        )

PCA_means_SCR

ggsave(path = "C:/Users/1aria/Desktop/rep/Output/Plots", filename = "PCA_patients_before_treatment.png",PCA_means_SCR,height=15,width=17,units="cm",dpi=200)
```

### Plot 2 preferred by Nidhi: all datapoints

```{r}
PCA_means_all_points <- ggplot(results_pca_means_N_NR, aes(x=PC1, y=PC2, color = Sample_status)) + 
  geom_point(size = 3)+ # alpha = as.integer(as.factor(Cycle))
  geom_vline(xintercept = 0, linetype=2, color = 'grey67') +
  geom_hline(yintercept = 0, linetype=2, color = 'grey67') +
  theme_bw() +
  xlim(c(-60, 60))+
  ylim(c(-50, 50))+
  labs(color= "Treatment outcome", title = 'All samples', subtite = 'Replicates have been averaged') +
  scale_color_manual(values=c('royalblue4','coral3'), labels = c("Non Responder", "Responder"))+
  theme(panel.grid.major = element_blank(), 
        panel.grid.minor = element_blank(),
        legend.position = 'right')

PCA_means_all_points
```

### Layout combined - bird's eye view

```{r}
layout1 <- ggarrange(PCA_means_all_points, PCA_w_repl_only, common.legend = TRUE, legend = "bottom")
layout1ann <- annotate_figure(layout1, top = text_grob("PCA plots",  color = "#333333", face = "bold", size = 15))

ggsave(path = "C:/Users/1aria/Desktop/rep/Output/Plots", filename = "PCA_all_and_means_panel.pdf",layout1ann,height=13,width=21,units="cm",dpi=200)
```

### Second layout: Clusters each infividuals

```{r}

results_pca_means_N_NR <- results_pca_means %>% filter(Sample_status %in% c('R', 'NR') )

PCA_means_clusters <- ggplot(results_pca_means_N_NR, aes(x=PC1, y=PC2, color = as.factor(Cycle))) + 
  geom_point(size = 3)+ # alpha = as.integer(as.factor(Cycle))
  geom_vline(xintercept = 0, linetype=2) +
  geom_hline(yintercept = 0, linetype=2) +
  theme_bw() +
  xlim(c(-60, 60))+
  ylim(c(-50, 50))+
  scale_color_manual()+
  labs(title = "PCA of each patient", color = "Treatment cycle") +
  scale_color_manual(labels = c("Screen", "C3", 'C4', 'C5', 'EoT'), values = bucket_cols)+
    theme(panel.grid.major = element_blank(), 
        panel.grid.minor = element_blank(),
        legend.position = 'bottom') +
  facet_wrap(~ Sample_nr)

PCA_means_clusters

ggsave(path = "C:/Users/1aria/Desktop/rep/Output/Plots", filename = "PCA_facet_each_patient.png",PCA_means_clusters,height=15,width=17,units="cm",dpi=200)

### Define here the gradient of cols you want: ###
bucket_cols <- c("0" = "#C7C7C7", "1" = "#919AB3", "2" = "#5C6D9F", "3" = "#27408B", '9' = 'black')

colfunc <- colorRampPalette(c("grey78", "royalblue4"))
colfunc(4)
plot(rep(1,4),col=colfunc(4),pch=19,cex=3)


```

### Second layout - plot 1 - Individual patient's Clusters (R)

```{r}
results_pca_means_R <- results_pca_means %>% filter(Sample_status %in% c('R') )


# Define here the gradient of cols you want: ###
bucket_cols_R <- c("0" = "#C7C7C7", '1' = '#C9A39B', '2' = 'coral3', '3' = 'coral4', '9' = 'black')   # 9 won't be used

colfunc <- colorRampPalette(c("coral3", "coral4"))
colfunc(5)
plot(rep(1,5),col=colfunc(5),pch=19,cex=3)

# PLot
PCA_means_clusters_R <- ggplot(results_pca_means_R, aes(x=PC1, y=PC2, color = as.factor(Cycle))) + 
  geom_point(size = 2)+ # alpha = as.integer(as.factor(Cycle))
  geom_vline(xintercept = 0, linetype=2, color = 'grey67') +
  geom_hline(yintercept = 0, linetype=2, color = 'grey67') +
  theme_bw() +
  xlim(c(-60, 30))+
  ylim(c(-20, 20))+
  labs(title = "Responders", color = "Treatment cycle") +
  scale_color_manual(labels = c("Screen", "C3", 'C4', 'C5'), values = bucket_cols_R)+
    theme(panel.grid.major = element_blank(), 
        panel.grid.minor = element_blank(),
        legend.position = 'bottom',
        axis.text.x = element_text(angle = 0, size = 8),
        strip.text = element_text(colour = 'white', face = 'bold'),
        strip.background =element_rect(fill="coral3")) +
  facet_wrap(~ Sample_nr)

PCA_means_clusters_R

ggsave(path = "C:/Users/1aria/Desktop/rep/Output/Plots", filename = "PCA_facet_each_R_patient.png",PCA_means_clusters_R,height=15,width=17,units="cm",dpi=200)


```

### Second layout - plot 2 - Individual patient's Clusters (NR)

```{r}
# df (do only once)
results_pca_means_NR <- results_pca_means %>% filter(Sample_status %in% c('NR') )

### Define here the gradient of cols you want: ###
bucket_cols_NR <- c("0" = "#919AB3", "1" = "#27408B", '9' = 'black')

colfunc <- colorRampPalette(c("#C7C7C7", "royalblue4"))
colfunc(4)
plot(rep(1,4),col=colfunc(4),pch=19,cex=3)

# plot
PCA_means_clusters_NR <- ggplot(results_pca_means_NR, aes(x=PC1, y=PC2, color = as.factor(Cycle))) + 
  geom_point(size = 2)+ # alpha = as.integer(as.factor(Cycle))
  geom_vline(xintercept = 0, linetype=2, color = 'grey67') +
  geom_hline(yintercept = 0, linetype=2, color = 'grey67') +
  theme_bw() +
  xlim(c(-60, 30))+
  ylim(c(-20, 20))+
  labs(title = "Non Responders", color = "Treatment cycle") +
  scale_color_manual(labels = c("Screen", "C3", 'EoT'), values = bucket_cols_NR)+
    theme(panel.grid.major = element_blank(), 
        panel.grid.minor = element_blank(),
        legend.position = 'bottom',
        axis.text.x = element_text(angle =0, size = 8),
        axis.text.y = element_blank(),
        axis.title.y = element_blank(),
        strip.text = element_text(colour = 'white', face = 'bold'),
        strip.background =element_rect(fill="royalblue4")) +
  facet_wrap(~ Sample_nr)

PCA_means_clusters_NR

ggsave(path = "C:/Users/1aria/Desktop/rep/Output/Plots", filename = "PCA_facet_each_NR_patient.png",PCA_means_clusters_NR,height=15,width=17,units="cm",dpi=200)



```

### Second layout - combine plots

```{r}
layout2 <- ggarrange(PCA_means_clusters_R, PCA_means_clusters_NR, common.legend = FALSE)

layout2ann <- annotate_figure(layout2, top = text_grob("PCA plots individual patients", color = "#333333", face = "bold", size = 15))

ggsave(path = "C:/Users/1aria/Desktop/rep/Output/Plots", filename = "PCA_individuals_NR_R_separated.pdf",layout2ann, height=10,width=21,units="cm",dpi=200)

```

# PCA by set

```{r}
results_pca_means_N_NR 

PCA_means_by_set <- ggplot(results_pca_means_N_NR, aes(x=PC1, y=PC2, color = as.factor(Set_nr))) + 
  geom_point(size = 3)+ # alpha = as.integer(as.factor(Cycle))
  geom_vline(xintercept = 0, linetype=2, color = 'grey67') +
  geom_hline(yintercept = 0, linetype=2, color = 'grey67') +
  theme_bw() +
  xlim(c(-60, 60))+
  ylim(c(-50, 50))+
  labs(color= "Set number", title = 'All samples', subtitle = 'Average of replicates') +
  # scale_color_manual(values=c('royalblue4','coral3', 'green'), labels = c("Non Responder", "Responder"))+
  theme(panel.grid.major = element_blank(), 
        panel.grid.minor = element_blank(),
        legend.position = 'right')

PCA_means_by_set

```

NB: Previously, we turned NAs into 0s (Alternative approaches I saw used are to infer NAs using BAyesian methods. Which should be more accurate.)

### Make a PCA object + scree plot

```{r}
# PCA object
wide_matrix_no_NAs_select <- wide_matrix_no_NAs %>% select(-c(1:9))
results.pca <-prcomp(wide_matrix_no_NAs_select[,],scale=TRUE,retx=TRUE)

# Scree plot
frac_var <- function(x) x^2/sum(x^2)
library(scales)

results.pca$sdev %>% 
  as_tibble() %>% 
  frac_var() %>% 
  mutate(Comp = colnames(results.pca$x)) %>% 
  slice(1:9) %>% 
  ggplot(aes(x=Comp, y = value)) + 
  geom_bar(stat = "identity", fill = "#4DC5F9") +
  geom_hline(yintercept = 0.03, linetype=2) +
  xlab("Principal Components") +
  scale_y_continuous(name = "Variance Explained", breaks = seq(0,0.8,0.1), labels = percent_format(accuracy = 5L)) +
  theme_classic(base_size = 14)
```

### Make a PCA plot

```{r}
results_pca_all <- results.pca$x %>%
  as.data.frame() %>%
  rownames_to_column(var = "Sample_name") %>%
  full_join(Metatable, by = "Sample_name")

ggplot(results_pca_all, aes(x=PC1, y=PC2, color = Sample_nr)) +
  geom_point(size = 3,  aes(shape = Cycle)) +
  geom_vline(xintercept = 0, linetype=2) +
  geom_hline(yintercept = 0, linetype=2) +
  theme_bw() +
  theme(panel.grid.major = element_blank(), 
        panel.grid.minor = element_blank())

plot(results.pca)

```
