==> R_scripts/04_Analysis_02_Volcano_plot.Rmd <==
---
title: "Make Volcano plot"
output:
  html_document:
    df_print: paged
editor_options:
  chunk_output_type: inline
fig_width: 15
fig_height: 17
---

### Admin

```{r, warning=F, results=F}
library(tidyverse)
library(readxl)
library(ggrepel)
library(tibble)
```

### Check if you have df in Data\>For_volcano.

If so, skip 0 and go directly to step 1. Else, proceed with creating relevant dfs

```{r}
load("C:/Users/1aria/Desktop/rep/Data/Output_data/For_Volcano/NR_vs_R_results.Rda")
```

### 0. Load dataframes

Use the one with means of replicate values, no batch1, no ICI

```{r}

load("C:/Users/1aria/Desktop/rep/Data/Output_data/wide_matrix_t_means.Rda")
Metatable_means <- read_excel("C:/Users/1aria/Desktop/rep/Data/Metadata/Metatable_means.xlsx", sheet = "w nur")

# remove batch 1
wide_matrix_t_means_no_batch1 <- wide_matrix_t_means %>% select(-contains(c('NR_13', 'NR_14', 'NR_15', 'NR_16', 'NR_17', 'NR_18')))
```

### 0. Define diff_expression function

```{r}
# Provide: 
# dataset: numerical only df with genes on rows and samples as col
# colnames_conditionA/B: list containing names of all cols of interest (eg NR vs R)

diff_exp = function(dataset, colnames_conditionA, colnames_conditionB) {
  results = apply(dataset, 1, function(x) t.test(x[colnames_conditionA], x[colnames_conditionB], na.action=na.omit))
  
  colindex_conditionA <-which(colnames(dataset) %in% colnames_conditionA)
  colindex_conditionB  <-which(colnames(dataset) %in% colnames_conditionB)
  
  
  p.val = list()
  for(i in names(results)) {p.val[[i]] = results[[i]]$p.value}
  p_value = as.vector(unlist(p.val))
  
  log2_fc = rowMeans(dataset[colindex_conditionA], na.rm=TRUE)-rowMeans(dataset[colindex_conditionB], na.rm=TRUE)
  
  t_test_result = as.data.frame(cbind(names(results), log2_fc, p_value))
  colnames(t_test_result) = c("Protein", "log2_fc", "p_value")
  
  t_test_result$adj_p_value = p.adjust(t_test_result$p_value, method="fdr")
  
  # t_test_result$Gene_Name = data$X[match(res$Protein, data$geneid)]
  
  t_test_result
}
```

### 0. Define paired t-test

```{r}
diff_exp_paired = function(dataset, colnames_conditionA, colnames_conditionB) {
  results = apply(dataset, 1, function(x) t.test(x[colnames_conditionA], x[colnames_conditionB], paired = TRUE, na.action=na.omit))
  
  colindex_conditionA <-which(colnames(dataset) %in% colnames_conditionA)
  colindex_conditionB  <-which(colnames(dataset) %in% colnames_conditionB)
  
  
  p.val = list()
  for(i in names(results)) {p.val[[i]] = results[[i]]$p.value}
  p_value = as.vector(unlist(p.val))
  
  log2_fc = rowMeans(dataset[colindex_conditionA], na.rm=TRUE)-rowMeans(dataset[colindex_conditionB], na.rm=TRUE)
  
  t_test_result = as.data.frame(cbind(names(results), log2_fc, p_value))
  colnames(t_test_result) = c("Protein", "log2_fc", "p_value")
  
  t_test_result$adj_p_value = p.adjust(t_test_result$p_value, method="fdr")   # fdr = Benj Hochberg
  
  t_test_result
}
```

# N vr NR

### 0. Diff expression of R vs NR

```{r}
# Define conditions -->  colnames of R vs NR
NR_colnames <- Metatable_means %>% filter(Sample_status == 'NR') %>% select(Sample_name) %>% as.list() %>% unlist()
R_colnames <- Metatable_means %>% filter(Sample_status == 'R') %>% select(Sample_name) %>% as.list() %>% unlist()

# Run function
NR_vs_R_results = as.data.frame(diff_exp(wide_matrix_t_means_no_batch1, colnames_conditionA =NR_colnames, colnames_conditionB =R_colnames ))

save(NR_vs_R_results, file = file.path('C:','Users', '1aria', 'Desktop', 'rep', 'Data', 'Output_data', 'For_Volcano','NR_vs_R_results.Rda'))

```

### 1. Subset df to find significant matches

```{r}
# Which proteins are diff expressed? (adj p below 0.05?)
significant_NR_vs_R = NR_vs_R_results %>% na.omit()  %>%  filter(adj_p_value<0.05)  # 19 proteins are signid different expressed



sign_NR_vs_R_results_expanded_DOWN_fc05 <- NR_vs_R_results_expanded %>% filter(diff_expressed %in% c('DOWN')) 
sign_NR_vs_R_results_expanded_UP_fc05 <- NR_vs_R_results_expanded %>% filter(diff_expressed %in% c('UP'))  
save(sign_NR_vs_R_results_expanded_DOWN_fc05, file = "C:/Users/1aria/Desktop/rep/Data/Output_data/For_GSEA/NR_vs_R_adj_DOWN_fc05.Rda")
save(sign_NR_vs_R_results_expanded_UP_fc05, file = "C:/Users/1aria/Desktop/rep/Data/Output_data/For_GSEA/NR_vs_R_adj_UP_fc05.Rda")

# Save
write.table(sign_NR_vs_R_results_expanded_DOWN$Protein, "C:/Users/1aria/Desktop/rep/Data/Output_data/Significantly_diff_expressed/NR_vs_R_adj_DOWN_fc0.5.txt", col.names = FALSE, row.names = FALSE, quote = FALSE)

write.table(sign_NR_vs_R_results_expanded_UP$Protein, "C:/Users/1aria/Desktop/rep/Data/Output_data/Significantly_diff_expressed/NR_vs_R_adj_UP_fc0.5.txt", col.names = FALSE, row.names = FALSE, quote = FALSE)
```

### 2. Plot the expanded df - FC:0.5

```{r}
# expand results so that easier to make volcano plot - add cols with additional info
NR_vs_R_results_expanded_0.5 <- NR_vs_R_results   %>% 
  na.omit() %>%
  mutate_at(c('log2_fc', 'p_value', 'adj_p_value'), as.numeric) %>%
  mutate(diff_expressed = "NO") %>%
  mutate(diff_expressed = if_else(condition = (log2_fc > 0.5 & adj_p_value < 0.05) , true = "UP", false = diff_expressed )) %>%
  mutate(diff_expressed = if_else(condition = (log2_fc < -0.5 & adj_p_value < 0.05) , true = "DOWN", false = diff_expressed ))   %>%
  mutate(label = NA ) %>%
  mutate(label = if_else(condition = (diff_expressed != "NO"), true = Protein, false = label))

save(NR_vs_R_results_expanded_0.5, file = 'NR_vs_R_results_expanded_0.5.Rda')


volc_NR_vs_R_FC05 <- ggplot(data=NR_vs_R_results_expanded_0.5, aes(x=log2_fc, y=-log10(adj_p_value))) + 
  geom_point(size = 2, aes(colour=diff_expressed)) +
  theme_bw() + 
  geom_vline(xintercept=c(-0.5, 0.5), col="grey60") +
  geom_hline(yintercept=-log10(0.05), col="grey60") +
  geom_vline(xintercept=c(0, 0), col="grey60",lty = 2 ) +
  scale_color_manual(values=c( "black", "coral3"), labels = c("Non significant", "Up-regulated")) +
  geom_text_repel(aes(label = label, color = diff_expressed), max.overlaps = Inf, show.legend = FALSE) +
  scale_x_continuous(breaks = c(-1.5, -1, -0.5, -0.3, 0, 0.3, 0.5, 1, 1.5), limits = c(-1.2, 1.2))+
  theme(panel.grid.major = element_blank(), 
        panel.grid.minor = element_blank(),
        legend.position = 'bottom',
        plot.title = element_text(size = 15, face = 'bold')) +
  labs(title= 'Volcano plot - Non responders vs responders', color = "Differentially expressed", x = 'log2(fold change)', y= '- log10(p-value)')

volc_NR_vs_R_FC05

ggsave(path = "C:/Users/1aria/Desktop/rep/Output/Plots", filename = "Volcano_NR_vs_R.png",volc_NR_vs_R_FC05,height=15,width=17,units="cm",dpi=200)
```

### Explore results where significance defined as *unadjusted* p_value \< 0.05 & FC = 1

```{r}
NR_vs_R_results_expanded_unadj <- NR_vs_R_results   %>% 
  na.omit() %>%
  mutate_at(c('log2_fc', 'p_value', 'adj_p_value'), as.numeric) %>%
  mutate(diff_expressed = "NO") %>%
  mutate(diff_expressed = if_else(condition = (log2_fc > 1 & p_value < 0.05) , true = "UP", false = diff_expressed )) %>%
  mutate(diff_expressed = if_else(condition = (log2_fc < -1 & p_value < 0.05) , true = "DOWN", false = diff_expressed ))   %>%
  mutate(label = NA ) %>%
  mutate(label = if_else(condition = (diff_expressed != "NO"), true = Protein, false = label))


save(NR_vs_R_results_expanded_unadj, file = 'NR_vs_R_results_expanded_unadj .Rda')

volc_unadj <- ggplot(data=NR_vs_R_results_expanded_unadj, aes(x=log2_fc, y=-log10(p_value))) + 
  geom_point(size = 2, aes(colour=diff_expressed)) +
  theme_bw() + 
  geom_vline(xintercept=c(-0.5, 0.5), col="grey60") +
  geom_hline(yintercept=-log10(0.05), col="grey60") +
  geom_vline(xintercept=c(0, 0), col="grey60",lty = 2 ) +
  scale_color_manual(values=c("royalblue4", "black", "coral3"), labels = c("Down-regulated", "Non significant", "Up-regulated")) +
  geom_text_repel(aes(label = label, color = diff_expressed), max.overlaps = Inf, show.legend =  FALSE) +
  scale_x_continuous(breaks = c(-1.5, -1, -0.5, 0, 0.5, 1, 1.5), limits = c(-1.2, 1.2))+
  theme(panel.grid.major = element_blank(), 
        panel.grid.minor = element_blank(),
        legend.position = 'bottom',
        plot.title = element_text(size = 15, face = 'bold')) +
  labs(title= 'Volcano plot - Non responders vs responders', subtitle = 'Unadjusted p-value', color = "Differentially expressed", x = 'log2(fold change)', y= '- log10(p-value (unadjusted))')

volc_unadj 

ggsave(path = "C:/Users/1aria/Desktop/rep/Output/Plots", filename = "Volcano_NR_vs_R_unadj.png",volc_unadj,height=15,width=17,units="cm",dpi=200)

# Find the up and downregulated proteins
sign_UP_NRvsR <- NR_vs_R_results_expanded_unadj %>% filter(diff_expressed %in% c('UP'))  %>%  select(Protein)
sign_DOWN_NRvsR <- NR_vs_R_results_expanded_unadj %>% filter(diff_expressed %in% c('DOWN'))  %>%  select(Protein)

# Save
write.table(sign_UP_NRvsR$Protein, "C:/Users/1aria/Desktop/rep/Data/Output_data/Significantly_diff_expressed/NR_vs_R_unadjp_upregul_fc1.txt", col.names = FALSE, row.names = FALSE, quote = FALSE)

write.table(sign_DOWN_NRvsR$Protein, "C:/Users/1aria/Desktop/rep/Data/Output_data/Significantly_diff_expressed/NR_vs_R_unadjp_downregul_fc1.txt", col.names = FALSE, row.names = FALSE, quote = FALSE)

```

### Explore results where significance defined as *unadjusted* p_value \< 0.05 & FC = 0.5

```{r}
NR_vs_R_results_expanded_unadj_fc0.5 <- NR_vs_R_results   %>% 
  na.omit() %>%
  mutate_at(c('log2_fc', 'p_value', 'adj_p_value'), as.numeric) %>%
  mutate(diff_expressed = "NO") %>%
  mutate(diff_expressed = if_else(condition = (log2_fc > 0.5 & p_value < 0.05) , true = "UP", false = diff_expressed )) %>%
  mutate(diff_expressed = if_else(condition = (log2_fc < -0.5 & p_value < 0.05) , true = "DOWN", false = diff_expressed ))   %>%
  mutate(label = NA ) %>%
  mutate(label = if_else(condition = (diff_expressed != "NO"), true = Protein, false = label))


# Find the UPregulated and downregulated proteins
sign_UP_NRvsR_fc05 <- NR_vs_R_results_expanded_unadj_fc0.5 %>% filter(diff_expressed %in% c('UP'))  %>%  select(Protein)

sign_DOWN_NRvsR_fc05 <- NR_vs_R_results_expanded_unadj_fc0.5 %>% filter(diff_expressed %in% c('DOWN'))  %>%  select(Protein)


#Save as lists
write.table(sign_UP_NRvsR_fc05$Protein, "C:/Users/1aria/Desktop/rep/Data/Output_data/Significantly_diff_expressed/NR_vs_R_unadjp_upregul_fc0.5.txt", col.names = FALSE, row.names = FALSE, quote = FALSE)

write.table(sign_DOWN_NRvsR_fc05$Protein, "C:/Users/1aria/Desktop/rep/Data/Output_data/Significantly_diff_expressed/NR_vs_R_unadjp_downregul_fc0.5.txt", col.names = FALSE, row.names = FALSE, quote = FALSE)

```

# NR vs R - more comparisons

No need to make volcano plots if no differecne found

## Compare only SCR points

No significant differences found

```{r}
# Define conditions -->  colnames of R vs NR
NR_colnames_SCR <- Metatable_means %>% filter(Sample_status == 'NR', Cycle == 0) %>% select(Sample_name) %>% as.list() %>% unlist()
R_colnames_SCR <- Metatable_means %>% filter(Sample_status == 'R', Cycle == 0) %>% select(Sample_name) %>% as.list() %>% unlist()

# Run function
SCR_NR_vs_R_results = as.data.frame(diff_exp(wide_matrix_t_means_no_batch1, colnames_conditionA =NR_colnames_SCR, colnames_conditionB =R_colnames_SCR ))

# Which proteins are diff expressed? (adj p below 0.05?)
significant_SCR_NR_vs_R = SCR_NR_vs_R_results %>% na.omit()  %>%  filter(adj_p_value<0.05)  # 0 proteins are signid different expressed
```

## Compare only C3

No significant differences found

```{r}
# Define conditions -->  colnames of R vs NR
NR_colnames_C3 <- Metatable_means %>% filter(Sample_status == 'NR', Cycle == 1) %>% select(Sample_name) %>% as.list() %>% unlist()
R_colnames_C3 <- Metatable_means %>% filter(Sample_status == 'R', Cycle == 1) %>% select(Sample_name) %>% as.list() %>% unlist()

# Run function
C3_NR_vs_R_results = as.data.frame(diff_exp(wide_matrix_t_means_no_batch1, colnames_conditionA =NR_colnames_C3, colnames_conditionB =R_colnames_C3 ))

# Which proteins are diff expressed? (adj p below 0.05?)
significant_C3_NR_vs_R = C3_NR_vs_R_results %>% na.omit()  %>%  filter(adj_p_value<0.05)  # 0 proteins are signid different expressed
```

## Compare SCR + C3

0 significant matches

```{r}
# NR_colnames_C3_SCR <- Metatable_means %>%
#   filter(Sample_status == c('NR'), Cycle == c(1,0)) %>%
#   select(Sample_name) %>%
#   as.list() %>% unlist()

NR_colnames_C3_SCR <- c(
  'NR_7_C3_Set1_tmt16plex_128C',
  'NR_8_SCR_Set1_tmt16plex_130C',
  'NR_9_C3_Set2_tmt16plex_130C',
  'NR_9_SCR_Set2_tmt16plex_mean',
  'NR_10_C3_Set3_tmt16plex_128C',
  'NR_7_SCR_Set1_tmt16plex_127C',
  'NR_8_C3_Set1_tmt16plex_131C',
  'NR_10_SCR_Set3_tmt16plex_127C',
  'NR_11_C3_Set3_tmt16plex_131C',
  'NR_11_SCR_Set3_tmt16plex_130C',
  'NR_12_C3_Set3_tmt16plex_133C',
  'NR_12_SCR_Set3_tmt16plex_132C'
)


R_colnames_C3_SCR <- Metatable_means %>%
  filter(Sample_status == 'R', Cycle == c(1, 0) ) %>%
  select(Sample_name) %>%
  as.list() %>% unlist()

# Run function
C3_SCR_NR_vs_R_results = as.data.frame(diff_exp(wide_matrix_t_means_no_batch1, colnames_conditionA =NR_colnames_C3_SCR, colnames_conditionB =R_colnames_C3_SCR ))

# Which proteins are diff expressed? (adj p below 0.05?)
significant_C3_SCR_NR_vs_R = C3_SCR_NR_vs_R_results %>% na.omit()  %>%  filter(adj_p_value<0.05)  # 0 proteins are signid different expressed
```

# Un-treated (SCR) vs Treated (C3, C4, C5, EoT) - all

```{r}
# Define conditions -->  colnames of R vs NR
untreated_colnames <- Metatable_means %>% filter(Cycle == '0', Sample_status %in% c( 'R')) %>% select(Sample_name) %>% as.list() %>% unlist()
treated_colnames <- Metatable_means %>% filter(Cycle != '0', Sample_status %in% c('R')) %>% select(Sample_name) %>% as.list() %>% unlist()

# Run function
Untreat_vs_treat_results_R = as.data.frame(diff_exp(wide_matrix_t_means_no_batch1, colnames_conditionA =untreated_colnames, colnames_conditionB =treated_colnames ))

save(Untreat_vs_treat_results, file = file.path('C:','Users', '1aria', 'Desktop', 'rep', 'Data', 'Output_data', 'For_Volcano','Untreat_vs_treat_results.Rda'))
```

# Un-treated (SCR) vs Treated (C3, C4, C5, EoT) - NR

```{r}
untreated_colnames_NR <- Metatable_means %>% filter(Cycle == '0', Sample_status %in% c( 'NR')) %>% select(Sample_name) %>% as.list() %>% unlist()
treated_colnames_NR <- Metatable_means %>% filter(Cycle != '0', Sample_status %in% c('NR')) %>% select(Sample_name) %>% as.list() %>% unlist()

# Run function
Untreat_vs_treat_results_NR = as.data.frame(diff_exp(wide_matrix_t_means_no_batch1, colnames_conditionA =untreated_colnames_NR, colnames_conditionB =treated_colnames_NR ))
```

### 1. Subset df to find significant matches

None found. Neither up nor down

```{r}
# Which proteins are diff expressed? (adj p below 0.05?)
significant_untre_vs_treat_R = Untreat_vs_treat_results_R %>% na.omit()  %>%  filter(adj_p_value<0.05)  # 0 proteins are signid different expressed


# expand - untreat vs treat (R)
Untreat_vs_treat_results_R_expanded <- Untreat_vs_treat_results_R   %>% 
  na.omit() %>%
  mutate_at(c('log2_fc', 'p_value', 'adj_p_value'), as.numeric) %>%
  mutate(diff_expressed = "NO") %>%
  mutate(diff_expressed = if_else(condition = (log2_fc > 0.3 & adj_p_value < 0.05) , true = "UP", false = diff_expressed )) %>%
  mutate(diff_expressed = if_else(condition = (log2_fc < -0.3 & adj_p_value < 0.05) , true = "DOWN", false = diff_expressed ))   %>%
  mutate(label = NA ) %>%
  mutate(label = if_else(condition = (diff_expressed != "NO"), true = Protein, false = label))

# expand - untreat vs treat (NR)
Untreat_vs_treat_results_NR_expanded <- Untreat_vs_treat_results_NR   %>% 
  na.omit() %>%
  mutate_at(c('log2_fc', 'p_value', 'adj_p_value'), as.numeric) %>%
  mutate(diff_expressed = "NO") %>%
  mutate(diff_expressed = if_else(condition = (log2_fc > 0.3 & adj_p_value < 0.05) , true = "UP", false = diff_expressed )) %>%
  mutate(diff_expressed = if_else(condition = (log2_fc < -0.3 & adj_p_value < 0.05) , true = "DOWN", false = diff_expressed ))   %>%
  mutate(label = NA ) %>%
  mutate(label = if_else(condition = (diff_expressed != "NO"), true = Protein, false = label))

                                                                
# Nr vs r
adj_p_NR_vs_R <- p.adjust(NR_vs_R_results_expanded$p_value, method="fdr")
plot(density(adj_p_NR_vs_R ))

#Treat vs untre (all)	
adj_p_untreat_vs_treat<- p.adjust(Untreat_vs_treat_results_expanded$p_value, method="fdr")
plot(density(adj_p_untreat_vs_treat))

#Treat vs untre (R)	
adj_p_untreat_vs_treat_R<- p.adjust(Untreat_vs_treat_results_R_expanded$p_value, method="fdr")
plot(density(adj_p_untreat_vs_treat_R))

#Treat vs untre (NR)	
adj_p_untreat_vs_treat_NR<- p.adjust(Untreat_vs_treat_results_NR_expanded$p_value, method="fdr")
plot(density(adj_p_untreat_vs_treat_NR))



save(Untreat_vs_treat_results_expanded, file = file.path('C:','Users', '1aria', 'Desktop', 'rep', 'Data', 'Output_data', 'For_Volcano','Untreat_vs_treat_results_expanded.Rda'))
```

### 2. Plot the expanded df

```{r}
# R
volc_untr_treat_R <- ggplot(data=Untreat_vs_treat_results_R_expanded, aes(x=log2_fc, y=-log10(adj_p_value))) + 
  geom_point(size = 2, aes(colour='coral3'))+
  theme_bw() + 
  geom_vline(xintercept=c(-0.3, 0.3), col="grey60") +
  geom_hline(yintercept=-log10(0.05), col="grey60") +
  geom_vline(xintercept=c(0, 0), col="grey60",lty = 2 ) +
  # scale_color_manual(values=c("royalblue4", "black", "coral3")) +
  geom_text_repel(aes(label = label, color = diff_expressed), max.overlaps = Inf, show.legend=FALSE) +
  scale_x_continuous(breaks = c(-1.5, -1, -0.5, -0.3, 0, 0.3, 0.5, 1, 1.5), limits = c(-1.2, 1.2))+
  scale_y_continuous(limits = c(0, 0.1))+
  guides(color = FALSE)+
  theme(panel.grid.major = element_blank(), 
        panel.grid.minor = element_blank(),
        legend.position = 'bottom',
        plot.title = element_text(size = 15, face = 'bold')) +
  labs(title= 'Volcano plot - Untreated (SCR) vs treated (C3, C4, C5, EoT), R', color = "Differentially expressed", x = 'log2(fold change)', y= '- log10(p-value)')

volc_untr_treat_R

# NR
volc_untr_treat_NR <- ggplot(data=Untreat_vs_treat_results_NR_expanded, aes(x=log2_fc, y=-log10(adj_p_value))) + 
  geom_point(size = 2, aes(colour=diff_expressed))+
  theme_bw() + 
  geom_vline(xintercept=c(-0.3, 0.3), col="grey60") +
  geom_hline(yintercept=-log10(0.05), col="grey60") +
  geom_vline(xintercept=c(0, 0), col="grey60",lty = 2 ) +
  scale_color_manual(values=c("royalblue4", "black", "coral3")) +
  geom_text_repel(aes(label = label, color = diff_expressed), max.overlaps = Inf, show.legend=FALSE) +
  scale_x_continuous(breaks = c(-1.5, -1, -0.5, -0.3, 0, 0.3, 0.5, 1, 1.5), limits = c(-1.2, 1.2))+
  scale_y_continuous(limits = c(0, 0.1))+
  guides(color = FALSE)+
  theme(panel.grid.major = element_blank(), 
        panel.grid.minor = element_blank(),
        legend.position = 'bottom',
        plot.title = element_text(size = 15, face = 'bold')) +
  labs(title= 'Volcano plot - Untreated (SCR) vs treated (C3, C4, C5, EoT), NR', color = "Differentially expressed", x = 'log2(fold change)', y= '- log10(p-value)')

volc_untr_treat_NR






# ggsave(path = "C:/Users/1aria/Desktop/rep/Output/Plots", filename = "Volcano_NR_vs_R.png",volc1,height=15,width=17,units="cm",dpi=200)
```

```{r}
plot(density(Untreat_vs_treat_results_expanded$adj_p_value))
```

### SCR vs EOT (NR)

```{r}
# Define conditions -->  colnames of SCR vs EOT (NR)
NR_colnames_SCR <- Metatable_means %>% filter(Sample_status == 'NR', Cycle == 0, Sample_nr %in% c(7, 8, 10, 11)) %>% select(Sample_name) %>% as.list()  %>%  unlist()
                                              
NR_colnames_EOT <- Metatable_means %>% filter(Sample_status == 'R', Cycle == 1) %>% select(Sample_name) %>% as.list()  %>%  unlist()

# Run function
SCRvsEOTForNR_results = as.data.frame(diff_exp(wide_matrix_t_means_no_batch1, colnames_conditionA =NR_colnames_SCR, colnames_conditionB =NR_colnames_EOT ))

# Manually
results = apply(wide_matrix_t_means_no_batch1, 1, function(x) t.test(x[NR_colnames_SCR], x[NR_colnames_EOT], paired = TRUE, na.action=na.omit))
  
  colindex_conditionA <-which(colnames(dataset) %in% colnames_conditionA)
  colindex_conditionB  <-which(colnames(dataset) %in% colnames_conditionB)
  
  
  p.val = list()
  for(i in names(results)) {p.val[[i]] = results[[i]]$p.value}
  p_value = as.vector(unlist(p.val))
  
  log2_fc = rowMeans(dataset[colindex_conditionA], na.rm=TRUE)-rowMeans(dataset[colindex_conditionB], na.rm=TRUE)
  
  t_test_result = as.data.frame(cbind(names(results), log2_fc, p_value))
  colnames(t_test_result) = c("Protein", "log2_fc", "p_value")
  
  t_test_result$adj_p_value = p.adjust(t_test_result$p_value, method="fdr")   # fdr = Benj Hochberg
  
  t_test_result


# Which proteins are diff expressed? (adj p below 0.05?)
significant_SCR_NR_vs_R = SCR_NR_vs_R_results %>% na.omit()  %>%  filter(adj_p_value<0.05)  # 0 proteins are signid different expressed
```
