==> R_scripts/04_Analysis_03_Heatmap.Rmd <==
---
title: "04 Analysis - 03 Heatmap"
output: html_notebook
editor_options: 
  chunk_output_type: inline
---

Load packages

```{r}
library(tidyverse)
library(ggpubr)
library(pheatmap)
library(readxl)
library(janitor)
library(styler)

```

# Start here - check data_output \> for_heatmap

If present, proceed from 1. If not, proceed from 0

```{r}

load("C:/Users/1aria/Desktop/rep/Data/Output_data/For_heatmap/annotations_heatmap.Rda")
load("C:/Users/1aria/Desktop/rep/Data/Output_data/For_heatmap/annotations_heatmap_t.Rda")
load("C:/Users/1aria/Desktop/rep/Data/Output_data/For_heatmap/ungappy_wide_matrix_t_means_renamed.Rda")

```

### 0 - Load df (created in the PCA script)

```{r}
load("C:/Users/1aria/Desktop/rep/Data/Output_data/For_PCA/ungappy_wide_matrix_t_for_PCA.Rda")
load("C:/Users/1aria/Desktop/rep/Data/Output_data/For_PCA/ungappy_wide_matrix_t_means_for_PCA.Rda")
Metatable_means <- read_excel("C:/Users/1aria/Desktop/rep/Data/Metadata/Metatable_means.xlsx", sheet = "w nur")

Metatable_means_R_NR <- Metatable_means %>% filter(Sample_status %in% c('NR', 'R'))
```

### 0 - Add metatable data

```{r}
ungappy_wide_matrix_means_annot <- as.data.frame(t(ungappy_wide_matrix_t_means))%>%
  rownames_to_column(var = 'Sample_name') %>%
  full_join(Metatable_means, by = "Sample_name")

ungappy_wide_matrix_t_means_annot <- ungappy_wide_matrix_means_annot %>%
  data.table::transpose(keep.names = 'Gene_name', make.names = 'Sample_name') %>%
  column_to_rownames(var = 'Gene_name')


```

### 0 - modify df for better readability and save for later use

```{r}
ungappy_wide_matrix_t_means_renamed <- ungappy_wide_matrix_t_means %>% 
  rename('7_C3' = colnames(ungappy_wide_matrix_t_means[1]),
         '9_C3' = colnames(ungappy_wide_matrix_t_means[2]),
         '10_C3' = colnames(ungappy_wide_matrix_t_means[3]),
         '11_C3' = colnames(ungappy_wide_matrix_t_means[4]),
         '12_C3' = colnames(ungappy_wide_matrix_t_means[5]),
         '7_EOT' = colnames(ungappy_wide_matrix_t_means[6]),
         '10_EOT' = colnames(ungappy_wide_matrix_t_means[7]),
         '7_SCR' = colnames(ungappy_wide_matrix_t_means[8]),
         '10_SCR' = colnames(ungappy_wide_matrix_t_means[9]),
         '11_SCR' = colnames(ungappy_wide_matrix_t_means[10]),
         '12_SCR' = colnames(ungappy_wide_matrix_t_means[11]),
         '8_C3' = colnames(ungappy_wide_matrix_t_means[12]),
         '8_EOT' = colnames(ungappy_wide_matrix_t_means[13]),
         '8_SCR' = colnames(ungappy_wide_matrix_t_means[14]),
         '2_C3' = colnames(ungappy_wide_matrix_t_means[15]),
         '3_C3' = colnames(ungappy_wide_matrix_t_means[16]),
         '4_C3' = colnames(ungappy_wide_matrix_t_means[17]),
         '5_C3' = colnames(ungappy_wide_matrix_t_means[18]),
         '6_C3' = colnames(ungappy_wide_matrix_t_means[19]),
         '3_C4' = colnames(ungappy_wide_matrix_t_means[20]),
         '4_C4' = colnames(ungappy_wide_matrix_t_means[21]),
         '6_C4' = colnames(ungappy_wide_matrix_t_means[22]),
         '3_C5' = colnames(ungappy_wide_matrix_t_means[23]),
         '4_C5' = colnames(ungappy_wide_matrix_t_means[24]),
         '6_C5' = colnames(ungappy_wide_matrix_t_means[25]),
         '3_SCR' = colnames(ungappy_wide_matrix_t_means[26]),
         '4_SCR' = colnames(ungappy_wide_matrix_t_means[27]),
         '5_SCR' = colnames(ungappy_wide_matrix_t_means[28]),
         '6_SCR' = colnames(ungappy_wide_matrix_t_means[29]),
         '1_C3' = colnames(ungappy_wide_matrix_t_means[30]),
         '1_C4' = colnames(ungappy_wide_matrix_t_means[31]),
         '1_C5' = colnames(ungappy_wide_matrix_t_means[32]),
         '1_SCR' = colnames(ungappy_wide_matrix_t_means[33]),
         '2_SCR' = colnames(ungappy_wide_matrix_t_means[34]),
         '9_SCR' = colnames(ungappy_wide_matrix_t_means[35]))
  
save(ungappy_wide_matrix_t_means_renamed, file = 'ungappy_wide_matrix_t_means_renamed.Rda')
                          
```

```         
```

```{r}
# Annotation from metatable
df_annotations <- Metatable_means %>% filter(Sample_status %in% c('R', 'NR')) %>% select(-c('Replicate'))


annotations_heatmap <-
  data.frame(cbind(names(df_annotations), t(df_annotations))) %>% select(-c('X1')) %>% row_to_names(row_number = 1) %>%
  rename(
    '1_C3' = 1,
    '1_C4' = 2,
    '1_C5' = 3,
    '1_SCR' = 4,
    '2_C3' = 5,
    '2_SCR' = 6,
    '3_C3' = 7,
    '3_C4' = 8,
    '3_C5' = 9,
    '3_SCR' = 10,
    '4_C3' = 11,
    '4_C4' =12,
    '4_C5' = 13,
    '4_SCR' = 14,
    '5_C3' = 15,
    '5_SCR' = 16,
    '6_C3' = 17,
    '6_C4' = 18,
    '6_C5' = 19,
    '6_SCR' = 20,
    '7_C3' = 21,
    '7_EOT' = 22,
    '7_SCR' = 23,
    '8_C3' = 24,
    '8_EOT' = 25,
    '8_SCR' = 26,
    '9_C3' = 27,
    '9_SCR' = 28,
    '10_C3' = 29,
    '10_EOT' = 30,
    '10_SCR' = 31,
    '11_C3' = 32,
    '11_SCR' = 33,
    '12_C3' = 34,
    '12_SCR' = 35
  )

save(annotations_heatmap, file = 'C:/Users/1aria/Desktop/rep/Data/Output_data/For_heatmap/annotations_heatmap.Rda')

annotations_heatmap_t <- data.frame(cbind(names(annotations_heatmap), t(annotations_heatmap))) %>% select(-c('V1'))

save(annotations_heatmap_t, file = 'C:/Users/1aria/Desktop/rep/Data/Output_data/For_heatmap/annotations_heatmap_t.Rda')

save(annotations_heatmap_t, file = '~Data/Output_data/For_heatmap/annotations_heatmap_t.Rda')
```

# 1. Make different heatmaps

### Heatmap1 - annotate the sample nr

```{r}
# pick annotations
anndf_cols <- anndf  %>% select(c('Sample_status', 'Sample_nr'))

# colours
my_colour = list(
    Sample_status = c(NR = "royalblue4", R = "coral3"),
    Sample_nr = bucket_cols
    )
    
    
bucket_cols <- c("0" = "#8B0A50", "1" = "#951848", "2" = "#A02641", "3" = "#AA343A", '4' = '#B54232',
                 '5' = '#BF502B', '6' = '#CA5E24', '7' = '#D46C1D', '8' = '#DF7A15', '9' = '#E9880E', '10' = '#F49607', '11' = '#FFA500', '12' =  'tan1')

colfunc <- colorRampPalette(c("deeppink4", "orange"))
colfunc(12)
plot(rep(1,12),col=colfunc(12),pch=19,cex=3)

#plot

heat1 <- pheatmap(ungappy_wide_matrix_t_means_renamed ,scale="row",
                  show_rownames=FALSE, 
                  annotation_col = anndf_cols,
                  annotation_colors = my_colour,
                  cutree_cols = 3,
                  cutree_rows = 2,
                  main = 'Heatmap')
heat1

save_pheatmap_png <- function(x, filename, width=1200, height=1000, res = 150) {
  png(filename, width = width, height = height, res = res)
  grid::grid.newpage()
  grid::grid.draw(x$gtable)
  dev.off()
}
 
save_pheatmap_png(heat1, "heat1.png")

```

### Heatmap 2 - annotate sex

```{r}
# pick annotations
anndf_cols2 <- anndf  %>% select(c('Sample_status', 'Sex'))

# colours
 my_colour2 <- list(
    Sample_status = c(NR = "royalblue4", R = "coral3"),
    Sex = c('1' = "lightblue", '2' = "pink")
    )
    
    
bucket_cols <- c("0" = "#8B0A50", "1" = "#951848", "2" = "#A02641", "3" = "#AA343A", '4' = '#B54232',
                 '5' = '#BF502B', '6' = '#CA5E24', '7' = '#D46C1D', '8' = '#DF7A15', '9' = '#E9880E', '10' = '#F49607', '11' = '#FFA500', '12' =  'tan1')

colfunc <- colorRampPalette(c("deeppink4", "orange"))
colfunc(12)
plot(rep(1,12),col=colfunc(12),pch=19,cex=3)

#Plot
heat2 <- pheatmap(ungappy_wide_matrix_t_means_renamed ,scale="row",show_rownames=FALSE, annotation_col = anndf_cols2, annotation_colors = my_colour2)
heat2



```

### Heatmap 3 - annotate cycle

```{r}
# pick annotations
anndf_cols3 <- anndf  %>% select(c('Sample_status', 'Cycle'))

# colours
 my_colour3 <- list(
    Sample_status = c(NR = "royalblue4", R = "coral3"),
    Cycle = c('0' = "#C7C7C7", '1' = "#B3889F", '2' = '#9F4977', '3' = '#8B0A50', '9' = 'black')
    )
    
    
colfunc <- colorRampPalette(c("deeppink4", "#C7C7C7"))
colfunc(4)
plot(rep(1,4),col=colfunc(4),pch=19,cex=3)

#Plot
heat3 <- pheatmap(ungappy_wide_matrix_t_means_renamed ,scale="row",show_rownames=FALSE, annotation_col = anndf_cols3, annotation_colors = my_colour3)
heat3

```

### Heatmap 4 - annotate set nr

```{r}
# pick annotations
anndf_cols4 <- annotations_heatmap_t  %>% select(c('Sample_status', 'Set_nr'))

# colours
 my_colour4 <- list(
    Sample_status = c(NR = "royalblue4", R = "coral3")
    # Set_nr = c('1' = "#B3889F", '2' = '#9F4977', '3' = '#8B0A50', '9' = 'black')
    )
    
    
colfunc <- colorRampPalette(c("deeppink4", "#C7C7C7"))
colfunc(4)
plot(rep(1,4),col=colfunc(4),pch=19,cex=3)

#Plot
heat4 <- pheatmap(ungappy_wide_matrix_t_means_renamed ,scale="row",show_rownames=FALSE, annotation_col = anndf_cols4, annotation_colors = my_colour4, clustering_method = 'ward.D2', clustering_distance_rows = "euclidean"  )
heat4

```

### Heatmap 5 - annotate cancer type

```{r}
anndf_cols5 <- annotations_heatmap_t  %>% select(c("Sample_status", "TMB_levels", "Cancer_short"))

my_colour5 <- list(
  Sample_status = c(NR = "royalblue4", R = "coral3"),
  TMB_levels = c(
    'Low' = "limegreen",
    'Intermediate' = '#4F8821',
    'High' = '#8C3E2F',
    'Very high' = '#8B0000'
  ),
  Cancer_short = c(
    'ACC' = 'sienna1',
    'GC' = 'lightseagreen',
    'CRC' = 'salmon4',
    'LUSC' = 'lightsteelblue',
    'EC' = 'lightpink2',
    'BC' = 'plum3',
    'AC' = 'khaki',
    'PPC' = 'tan'
  )
)
    

colfunc <- colorRampPalette(c("darkred", "limegreen"))
colfunc(4)
plot(rep(1,4),col=colfunc(4),pch=19,cex=3)


heat5 <- pheatmap(ungappy_wide_matrix_t_means_renamed ,scale="row",show_rownames=FALSE, annotation_col = anndf_cols5, annotation_colors = my_colour5, height = 20, clustering_method = 'average')  # average = UPGMA = most commonly used method
heat5


# ggsave(path = "C:/Users/1aria/Desktop/rep/Output/Plots",   filename = "Heatmap5_average.png",heat5,height=20,width=17,units="cm",dpi=200)
```

### Try clustering

```{r}
clustering1 <- pheatmap(ungappy_wide_matrix_t_means_renamed ,scale="row",show_rownames=FALSE, annotation_col = anndf_cols5, annotation_colors = my_colour5, height = 20, cluster_rows = TRUE)
clustering1
```

# Find gene clusters

### Dendogram clustering of rows

```{r}
my_hclust_gene <- hclust(dist(ungappy_wide_matrix_t_means_renamed), method = "average")

as.dendrogram(my_hclust_gene)%>%
  plot(horiz = TRUE)

my_gene_col <- cutree(my_hclust_gene, k = 2)
my_gene_col <- data.frame(cluster = ifelse(test = my_gene_col == 1, yes = "cluster 1", no = "cluster 2"))


my_colour_clust <- list(
  Sample_status = c(NR = "royalblue4", R = "coral3"),
  Sex = c('1' = "#8B0A50", '2' = "orange"),
  cluster = c('cluster 1' = 'violet', 'cluster 2' = 'yellow')
  )

pheatmap(ungappy_wide_matrix_t_means_renamed ,scale="row",show_rownames=FALSE, 
         annotation_row = my_gene_col,
         annotation_colors = my_colour_clust)


```

### Cluster 12_C to see very red region

```{r}
Twelve_C3 <- ungappy_wide_matrix_t_means_renamed %>% select('12_C3')

distance_12C3 <- hclust(dist(Twelve_C3), method = "average")

as.dendrogram(distance_12C3) 
  plot(horiz = TRUE)

five_clusters_12C3 <- cutree(distance_12C3, k = 2) # make 5 clusters based on avg distance. List
Twelve_C3$cluster <- cluster_12C3 

clust_12_C3_heat <- pheatmap(Twelve_C3, scale="row",show_rownames=FALSE)
clust_12_C3_heat
```

```{r}

distance_12C3 <- hclust(dist(Twelve_C3), method = "average")

# in case u wanna see it
# as.dendrogram(distance_12C3) 
#   plot(horiz = TRUE)
  
five_clusters_12C3 <- cutree(distance_12C3, k = 5) # list of 5 clusters

sample_12c3_col <- data.frame(cluster = five_clusters_12C3)

my_colour_clust <- list(
  Sample_status = c(NR = "royalblue4", R = "coral3"),
  cluster = c('1' = "#8B0A50",
              '2' = "orange",
              "3" = 'red',
              '4' = 'yellow',
              '5' = 'green')
  )

pheatmap(ungappy_wide_matrix_t_means_renamed,scale="row",show_rownames=FALSE, 
         annotation_row =sample_12c3_col,
         annotation_colors = my_colour_clust)

```

# Find best clustering, based on diff linkage methods - use cluster::agnes
### Which linkage methods to use? One that maximizes agglomerative coefficient. 
```{r}
ungappy_wide_matrix_t_means_renamed_scaled <- scale(ungappy_wide_matrix_t_means_renamed)

m <- c( "average", "single", "complete", "ward", 'weighted')
names(m) <- c( "average", "single", "complete", "ward", 'weighted')

#function to compute agglomerative coefficient
agglomerative_coefficient <- function(m) {
  agnes(ungappy_wide_matrix_t_means_renamed, method = m)$ac
}

#calculate agglomerative coefficient for each clustering linkage method
sapply(m, agglomerative_coefficient)     # avg: 0.88, single: 0.81, complete: 0.91, ward: 0.96, weighted: 0.88

```
### Perform hierarc clusters
```{r}
clust <- agnes(ungappy_wide_matrix_t_means_renamed_scaled, method = "ward")
pltree(clust, cex = 0.6, hang = -1, main = "Dendrogram") 

```

### Gap stat for diff number of clusters
Here: no clutsers is best. 
```{r}
#calculate gap statistic for each number of clusters (up to 10 clusters)
gap_stat <- clusGap(ungappy_wide_matrix_t_means_renamed_scaled, FUN =  kmeans, nstart = 25, K.max = 7, B = 500)   #calc to up 10 clusters. B: nr montecarlo bootstrap samples

#produce plot of clusters vs. gap statistic
gap<- fviz_gap_stat(gap_stat) + theme(plot.title = element_text(size = 15, face = "bold", colour = "#333333", hjust = 0.5))
```
### pheatmap
```{r}
heat7 <- pheatmap(ungappy_wide_matrix_t_means_renamed ,scale="row",show_rownames=FALSE, annotation_col = anndf_cols5, annotation_colors = my_colour5, height = 20, clustering_method = 'ward.D2') 
heat7

ggsave(path = "C:/Users/1aria/Desktop/rep/Output/Plots", filename = "Heatmap_ward.png",heat7,dpi=200)
```
### Just to try: eliminate: 7_EOT, 1_C3, 1_C5, 12_C3
```{r}
ungappy_wide_matrix_t_means_renamed %>% select(-c('7_EOT', '1_C3', '1_C5', '12_C3')) %>% as.matrix() %>%
  heatmap(. ,scale="row",show_rownames=FALSE, height = 20, clustering_method = 'ward.D2') 
```

### heatmap of DE only
```{r}
DE_proteins <- NR_vs_R_results %>% filter(adj_p_value < 0.05) %>% select(Protein)

DE_ungappy <- ungappy_wide_matrix_t_means_renamed %>% filter(row.names(.) %in% DE_proteins$Protein)

anndf_cols8 <- annotations_heatmap_t  %>% select(c("Sample_status", "TMB_levels", "Cancer_short"))

heat8 <- pheatmap(DE_ungappy ,scale="row",show_rownames=TRUE, annotation_col = anndf_cols8, annotation_colors = my_colour5, height = 20, clustering_method = 'ward.D2') 

# linkage method? --> ward
m <- c( "average", "single", "complete", "ward", 'weighted')
names(m) <- c( "average", "single", "complete", "ward", 'weighted')

agglomerative_coefficient <- function(m) {
  agnes(DE_ungappy, method = m)$ac
}

sapply(m, agglomerative_coefficient)     # avg: 0.65, single: 0.421, complete: 0.74, ward: 0.81

# clusters?
gap_stat <- clusGap(DE_ungappy, FUN =  kmeans, nstart = 25, K.max = 12, B = 500)   #calc to up 10 clusters. B: nr montecarlo bootstrap samples

#produce plot of clusters vs. gap statistic
fviz_gap_stat(gap_stat) + theme(plot.title = element_text(size = 15, face = "bold", colour = "#333333", hjust = 0.5))

```
