==> R_scripts/04_Analysis_11_GSEA.Rmd <==
---
title: "GSAE"
output: html_notebook
editor_options: 
  chunk_output_type: inline
---
### Load packages

```{r}
BiocManager::install("AnnotationDbi")
library(clusterProfiler)
library(org.Hs.eg.db)
library(AnnotationDbi)
library(enrichplot)
library(dplyr)
library(ggplot2)
library(ggrepel)
library(forcats)
```

### Functions for own plots

```{r}
create_dotplot <- function(dataset, set.FDR.q.val, plot_name, title) {
  
  # Select only relevant columns for barplot
  genesets_upregulated <- dataset %>%
    dplyr::select(c("NAME", "SIZE", "NOM.p.val", "FDR.q.val")) 
  
  # Top gene sets - take only if FDR is below given threshold
  top_genesets_upregulated <- genesets_upregulated %>%
    arrange(FDR.q.val) %>%
    filter(FDR.q.val < set.FDR.q.val) %>%
    mutate(NAME = substring(NAME, first = 6))
  
  # Create dotplot
  ggplot(top_genesets_upregulated, aes(x = FDR.q.val, y = reorder(NAME, -FDR.q.val))) + 
    geom_point(aes(size = SIZE, color = '#1380A1'))+
    labs(title = title, x = 'p-value (FDR adjusted)', size = 'Nr of genes in set', y = 'KEGG pathway term')+
    guides(color = FALSE)+
    theme_bw()+
    theme(plot.title = element_text(size = 15, face = 'bold', hjust = 0.5))
  
}


create_meaningful_dotplot <- function(table_path, set.FDR.q.val, my_title) {
  
  # Insert table
  df <- read.table(table_path, sep='\t', header = TRUE)
  
  # Select only relevant columns for dotplot. Top 10
  genesets_upregulated <- df %>%
      dplyr::select(c("NAME", "SIZE", 'NES',"NOM.p.val", "FDR.q.val"))  %>%
      mutate(NAME = substring(NAME, first = 6)) %>%
      mutate(good_FDR = case_when(
      FDR.q.val < set.FDR.q.val ~ "YES",
      FDR.q.val > set.FDR.q.val ~ "NO"
      )) %>%
      slice(1:10)
  
  # Create dotplot
  my_plot <- ggplot(genesets_upregulated, aes(x = NES, y = reorder(NAME, NES))) + 
      geom_point(aes(size = SIZE, color = good_FDR == 'YES'))+
      labs(title = my_title, x = 'Normalized enrichment score (NES)', size = 'Nr of genes in set', y = 'KEGG pathway term')+
      scale_color_manual(name = 'FDR < 25%', values= setNames(c("coral3", "grey"), c(T, F))) +
      theme_bw()+
      guides(size = guide_legend(nrow = 2))+
      theme(plot.title = element_text(size = 12, face = 'bold', hjust = 0.5),
            legend.position="bottom",
            # legend.direction="horizontal",
            axis.text.y = element_text(size = 8),
            legend.justification = c("left", "bottom"))
  
  
my_plot
  
}

create_barplot_Nidhi <- function(table_path, slice_size, set.FDR.q.val, my_title) {
  
  # Insert table
  df <- read.table(table_path, sep='\t', header = TRUE)
  
  # Select only relevant columns for dotplot. Top 10
  genesets_upregulated <- df %>%
      dplyr::select(c("NAME", "SIZE", "NOM.p.val", "FDR.q.val", 'NES'))  %>%
      filter(NOM.p.val > 0.05) %>%
      mutate(NAME = substring(NAME, first = 6)) %>%
      mutate(good_FDR = case_when(
      FDR.q.val < set.FDR.q.val ~ "YES",
      FDR.q.val > set.FDR.q.val ~ "NO"
      )) %>%
      arrange('NES') %>%
      slice(1:slice_size)
  
  # Create dotplot
  my_plot_name <- ggplot(genesets_upregulated, aes(x = SIZE, y = reorder(NAME, SIZE))) + 
      geom_bar(stat = "identity", aes(fill = NOM.p.val))+
      labs(title = my_title, x = 'Genes in pathway', color = 'p-value', y = 'KEGG pathway term')+
      theme_bw()+
      labs(fill = 'P-value')+
      theme(plot.title = element_text(size = 12, face = 'bold', hjust = 0.5),
            legend.position="bottom",
            legend.direction="horizontal")
  
  my_plot_name
  
}
```


# GSEA - SCR (NR vs R)

### Make list
```{r}
load("C:/Users/1aria/Desktop/rep/Data/Output_data/Diff_expressed/ttest_NR_vs_R_means_SCR_no_batch1.Rda") #NR_vs_R_SCR_results
x <- as.data.frame(org.Hs.egSYMBOL2EG)


genes_up_NR_at_SCR <- NR_vs_R_SCR_results %>%
  mutate(Protein = gsub("\\.", "-",Protein)) %>%
  left_join(., x, by = c("Protein" = "symbol"))  %>% 
  mutate_at(c('log2_fc', 'p_value'), as.numeric) %>%
  # filter(log2_fc > 0) %>% 
  dplyr::select(gene_id, log2_fc) %>% arrange(., desc(log2_fc)) 

# remove where gene ID is na
geneList2 <-  genes_up_NR_at_SCR[,2] %>% as.numeric()
names(geneList2) = as.character(genes_up_NR_at_SCR[,1])
geneList2 = na.omit(geneList2)
geneList2 = sort(geneList2, decreasing = TRUE)
```

### GO
```{r}

############# GSEA of GO terms ################################
SCR_BP <- gseGO(geneList     = geneList2,
              OrgDb        = org.Hs.eg.db,
              ont          = "BP",
              minGSSize    = 10,
              maxGSSize    = 500,
              pvalueCutoff = 0.05,
              verbose      = FALSE)
SCR_MF<- gseGO(geneList     = geneList2,
              OrgDb        = org.Hs.eg.db,
              ont          = "MF",
              minGSSize    = 10,
              maxGSSize    = 500,
              pvalueCutoff = 0.05,
              verbose      = FALSE)
SCR_CC <- gseGO(geneList     = geneList2,
              OrgDb        = org.Hs.eg.db,
              ont          = "CC",
              minGSSize    = 10,
              maxGSSize    = 500,
              pvalueCutoff = 0.05,
              verbose      = FALSE)

SCR_BP_symbols <- setReadable(SCR_BP, 'org.Hs.eg.db', 'ENTREZID')
SCR_BP_symbols  <- as.data.frame(SCR_BP_symbols )
save(SCR_BP_symbols , file="C:/Users/1aria/Desktop/rep/Data/Output_data/GSEA_output/SCR_BP.Rda")

SCR_MF_symbols <- setReadable(SCR_MF, 'org.Hs.eg.db', 'ENTREZID')
SCR_MF_symbols  <- as.data.frame(SCR_MF_symbols )
save(SCR_MF_symbols , file="C:/Users/1aria/Desktop/rep/Data/Output_data/GSEA_output/SCR_MF.Rda")

SCR_CC_symbols <- setReadable(SCR_CC, 'org.Hs.eg.db', 'ENTREZID')
SCR_CC_symbols  <- as.data.frame(SCR_CC_symbols )
save(SCR_CC_symbols , file="C:/Users/1aria/Desktop/rep/Data/Output_data/GSEA_output/SCR_CC.Rda")

### plots ####

SCR_BP_pl<- dotplot(SCR_BP, showCategory = 10) + theme(axis.text.y = element_text(size = 9)) + ggtitle('GSEA of GO terms (BP) - SCR')+
   theme(plot.title = element_text(size = 12, face = 'bold', hjust = 0.5))

SCR_CC_pl<- dotplot(SCR_CC, showCategory = 10) + theme(axis.text.y = element_text(size = 9)) + ggtitle('GSEA of GO terms (CC) - SCR')+
   theme(plot.title = element_text(size = 12, face = 'bold', hjust = 0.5))

SCR_MF_pl<- dotplot(SCR_MF, showCategory = 10) + theme(axis.text.y = element_text(size = 9)) + ggtitle('GSEA of GO terms (MF) - SCR')+
   theme(plot.title = element_text(size = 12, face = 'bold', hjust = 0.5))

cnetplot(GSEA_GO_BP, categorySize="pvalue", foldChange=geneList) 

SCR_BP_pl
SCR_MF_pl
SCR_CC_pl

ggsave(path = "C:/Users/1aria/Desktop/rep/Output/Plots", filename = "SCR_BP_pl.png",SCR_BP_pl)
ggsave(path = "C:/Users/1aria/Desktop/rep/Output/Plots", filename = "SCR_CC_pl.png",SCR_CC_pl)
ggsave(path = "C:/Users/1aria/Desktop/rep/Output/Plots", filename = "SCR_MF_pl.png",SCR_MF_pl)

### Additional plots
SCR_BP_pairwise <- pairwise_termsim(SCR_BP )
SCR_MF_pairwise <- pairwise_termsim(SCR_MF)
SCR_CC_pairwise <- pairwise_termsim(SCR_CC)

tree_SCR_BP <- treeplot(SCR_BP_pairwise)
tree_SCR_MF <-treeplot(SCR_MF_pairwise)
tree_SCR_CC <-treeplot(SCR_CC_pairwise)

tree_SCR_BP

ggsave(path = "C:/Users/1aria/Desktop/rep/Output/Plots", filename = "tree_SCR_BP.png",tree_SCR_BP, width = 14, dpi=200)
ggsave(path = "C:/Users/1aria/Desktop/rep/Output/Plots", filename = "tree_SCR_MF.png",tree_SCR_MF, width = 14, dpi=200)
ggsave(path = "C:/Users/1aria/Desktop/rep/Output/Plots", filename = "tree_SCR_CC.png",tree_SCR_CC, width = 14, dpi=200)

enrichmap_C3_BP<-emapplot(SCR_BP_pairwise)

ggsave(path = "C:/Users/1aria/Desktop/rep/Output/Plots", filename = "enrichmap_C3_BP.png",enrichmap_C3_BP)

```

### KEGG
```{r}
GSEA_kegg2 <- gseKEGG(geneList     = geneList2,
               organism     = 'hsa',
               minGSSize    = 10,
               pvalueCutoff = 0.31,
               verbose      = FALSE)
kegg2.df <- as.data.frame(GSEA_kegg2)

# Turn kegg genes names from ENTREZ to symbols
GSEA_kegg_symbols2 <- setReadable(GSEA_kegg2, 'org.Hs.eg.db', 'ENTREZID')
kegg_symbols.df2 <- as.data.frame(GSEA_kegg_symbols2)


KEGG_SCR<- dotplot(GSEA_kegg2, showCategory=10) + ggtitle("KEGG GSEA - dysregulated at SCR") # works

ggsave(path = "C:/Users/1aria/Desktop/rep/Output/Plots", filename = "KEGG_SCR.png", KEGG_SCR, dpi = 200)

# network plot
kegg_netplot<- cnetplot(GSEA_kegg_symbols2, categorySize="p.adjust", foldChange=geneList2, circular = FALSE,
                        cex_gene = 0.5, cex_label_gene = 0.5) 
kegg_netplot
ggsave(path = "C:/Users/1aria/Desktop/rep/Output/Plots", filename = "KEGG_netplot.png",kegg_netplot)


heatplot(GSEA_kegg_symbols, foldChange=geneList, showCategory=10)

# PLots that require termwise networks (?) - enrichplot
kegg2 <- pairwise_termsim(GSEA_kegg_symbols)

treeplot(kegg2)
emapplot(kegg2)   # layout = 'kk'

```

# GSEA - C3 (NR vs R)
### Gene list
```{r}

### Create own geneList. (NB. need to convert HUGO symbol to NCBI gene ID) ####
load("C:/Users/1aria/Desktop/rep/Data/Output_data/Diff_expressed/ttest_NR_vs_R_means_C3_no_batch1.Rda")

x <- as.data.frame(org.Hs.egSYMBOL2EG)  # map HUGO to NCBI IDs

genes_up_NR_at_C3 <- NR_vs_R_C3_results %>%
  mutate(Protein = gsub("\\.", "-",Protein)) %>%
  left_join(., x, by = c("Protein" = "symbol"))  %>% 
  mutate_at(c('log2_fc', 'p_value'), as.numeric) %>%
  # filter(log2_fc > 0) %>% 
  dplyr::select(gene_id, log2_fc) %>% arrange(., desc(log2_fc)) 

# remove where gene ID is na
geneList1 <-  genes_up_NR_at_C3[,2] %>% as.numeric()
names(geneList1) = as.character(genes_up_NR_at_C3[,1])
geneList1 = na.omit(geneList1)
geneList1 = sort(geneList1, decreasing = TRUE)

#geneList contains 1787 named numbers. Name is the gene, number is the rank when no all log3´2FC


############# GSEA of GO terms ################################
GSEA_GO_BP_try <- gseGO(geneList     = geneList1,
              OrgDb        = org.Hs.eg.db,
              ont          = "BP",
              minGSSize    = 10,
              maxGSSize    = 500,
              pvalueCutoff = 0.05,
              verbose      = FALSE)

GSEA_GO_MF_C3<- gseGO(geneList     = geneList1,
              OrgDb        = org.Hs.eg.db,
              ont          = "MF",
              minGSSize    = 10,
              maxGSSize    = 500,
              pvalueCutoff = 0.05,
              verbose      = FALSE)

GSEA_GO_CC_C3 <- gseGO(geneList     = geneList1,
              OrgDb        = org.Hs.eg.db,
              ont          = "CC",
              minGSSize    = 10,
              maxGSSize    = 500,
              pvalueCutoff = 0.05,
              verbose      = FALSE)

GSEA_GO_BP_df_symbols <- setReadable(GSEA_GO_BP_try, 'org.Hs.eg.db', 'ENTREZID')
GSEA_GO_BP_df_symbols <- as.data.frame(GSEA_GO_BP_df_symbols)
save(GSEA_GO_BP_df_symbols, file="C:/Users/1aria/Desktop/rep/Data/Output_data/GSEA_output/C3_BP.Rda")

GSEA_GO_MF_C3_df_symbols <- setReadable(GSEA_GO_MF_C3, 'org.Hs.eg.db', 'ENTREZID')
GSEA_GO_MF_C3_df_symbols <- as.data.frame(GSEA_GO_MF_C3_df_symbols)
save(GSEA_GO_MF_C3_df_symbols, file="C:/Users/1aria/Desktop/rep/Data/Output_data/GSEA_output/C3_MF.Rda")

GSEA_GO_CC_C3_df_symbols <- setReadable(GSEA_GO_CC_C3, 'org.Hs.eg.db', 'ENTREZID')
GSEA_GO_CC_C3_df_symbols <- as.data.frame(GSEA_GO_CC_C3_df_symbols)
save(GSEA_GO_CC_C3_df_symbols, file="C:/Users/1aria/Desktop/rep/Data/Output_data/GSEA_output/C3_CC.Rda")

### plots ####

def_dotplot<- dotplot(GSEA_GO_BP, showCategory = 10) + theme(axis.text.y = element_text(size = 9)) + ggtitle('GSEA of GO terms (BP) - C3 dysregulated proteins')+
   theme(plot.title = element_text(size = 12, face = 'bold', hjust = 0.5))

def_dotplot 

C3_CC<- dotplot(GSEA_GO_CC_C3, showCategory = 10) + theme(axis.text.y = element_text(size = 9)) + ggtitle('GSEA of GO terms (CC) - C3')+
   theme(plot.title = element_text(size = 12, face = 'bold', hjust = 0.5))

C3_MF<- dotplot(GSEA_GO_MF_C3, showCategory = 10) + theme(axis.text.y = element_text(size = 9)) + ggtitle('GSEA of GO terms (MF) - C3')+
   theme(plot.title = element_text(size = 12, face = 'bold', hjust = 0.5))

cnetplot(GSEA_GO_BP, categorySize="pvalue", foldChange=geneList) 


C3_CC
C3_MF

ggsave(path = "C:/Users/1aria/Desktop/rep/Output/Plots", filename = "C3_CC.png",C3_CC)
ggsave(path = "C:/Users/1aria/Desktop/rep/Output/Plots", filename = "C3_MF.png",C3_MF)

### Additional plots
C3_BP_pairwise <- pairwise_termsim(GSEA_GO_BP_try )
C3_MF_pairwise <- pairwise_termsim(GSEA_GO_MF_C3)
C3_CC_pairwise <- pairwise_termsim(GSEA_GO_CC_C3)

tree_C3_BP <- treeplot(C3_BP_pairwise)
tree_C3_MF <-treeplot(C3_MF_pairwise)
tree_C3_CC <-treeplot(C3_CC_pairwise)

ggsave(path = "C:/Users/1aria/Desktop/rep/Output/Plots", filename = "tree_C3_BP.png",tree_C3_BP)
ggsave(path = "C:/Users/1aria/Desktop/rep/Output/Plots", filename = "tree_C3_MF.png",tree_C3_MF)
ggsave(path = "C:/Users/1aria/Desktop/rep/Output/Plots", filename = "tree_C3_CC.png",tree_C3_CC)

enrichmap_C3_BP<-emapplot(C3_BP_pairwise)

ggsave(path = "C:/Users/1aria/Desktop/rep/Output/Plots", filename = "enrichmap_C3_BP.png",enrichmap_C3_BP)

```
### KEGG

```{r}
GSEA_kegg1 <- gseKEGG(geneList     = geneList1,
               organism     = 'hsa',
               minGSSize    = 10,
               pvalueCutoff = 0.05,
               verbose      = FALSE)

kegg1.df <- as.data.frame(GSEA_kegg1)

# Turn kegg genes names from ENTREZ to symbols
GSEA_kegg1_symbols <- setReadable(GSEA_kegg1, 'org.Hs.eg.db', 'ENTREZID')
kegg1_symbols.df <- as.data.frame(GSEA_kegg1_symbols)

KEGG_C3 <- dotplot(GSEA_kegg1, showCategory=10) + ggtitle("KEGG GSEA - dysregulated at C3") 
KEGG_C3

ggsave(path = "C:/Users/1aria/Desktop/rep/Output/Plots", filename = "KEGG_C3.png", KEGG_C3, dpi = 200)

# save(kegg_symbols.df, file = "C:/Users/1aria/Desktop/rep/Data/clusterProfilR_GSEA/GSEA_kegg_symbold.Rda")

```

### NB: Plots available on clustrprofilr

```{r}

kegg_netplot<- cnetplot(GSEA_kegg1_symbols, categorySize="p.adjust", foldChange=geneList1, circular = FALSE,
                        cex_gene = 0.5, cex_label_gene = 0.5) 

heatplot(GSEA_kegg1_symbols, foldChange=geneList1, showCategory=10)

# PLots that require termwise networks
kegg2 <- pairwise_termsim(GSEA_kegg_symbols)

treeplot(kegg2)
emapplot(kegg2)   # layout = 'kk'
```
